backend/requirements.txt
# Example requirements.txt
django
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
django-cors-headers